exec: Enable basic functionality exec command
This patch enable basic functionality for exec command.

The runtime uses ssh to run new commands,
this require:
clear-container.img with openssh-server bundle

To be able to execute a command in the container
this execute:

	ssh -t root@container_ip command args

The IP address is  stored in the state file to know
where to connect after the container is created.

If a container does not have network
(expose an IP to connect), the exec command wont work.

- runtime  <cmd>  is optional when --process is used

- Adds struct to handle save process.json , process.json
nodes are not fully compatible with exsting structs, added:
struct cc_oci_process_exec

Signed-off-by: Jose Carlos Venegas Munoz <jose.carlos.venegas.munoz@intel.com>