pod: Track container mapping to pods
We add a pod configuration structure to cc_oci_config.
This allows us to track if a container either is a pod
sandbox, a regular pod container, or simply not belonging
to a pod at all.
With this mapping we can map a container creation request
(cor create container) to either a pod vm creation or to
an hyperstart command for creating a new container within
a pod managed by hyperstart.

As the OCI specification does not define pod specific
settings, we currently build the cc_pod structure from
OCI annotions. Currently only CRI-O builds these annotations.

Signed-off-by: Samuel Ortiz <sameo@linux.intel.com>