state: Track the container mapping to pods
We now track the container mapping to pod as
an oci_state pointer. It is rebuild from the OCI
JSON config file.

This allows us to not start the CRI-O pause container
needed to hold all namespace based pod containers, as
the pod VM will do so for Clear Containers.

Signed-off-by: Samuel Ortiz <sameo@linux.intel.com>