Process: Ignore SIGPIPE when running hooks.
If a hook process is not expecting to be passed state on its stdin and
exits before the runtime has had a chance to pass the state to the
child, the runtime will be sent SIGPIPE, causing it to be terminated.

To avoid this, disable SIGPIPE when running the hooks.

Signed-off-by: James O. D. Hunt <james.o.hunt@intel.com>