cor: proxy: Fix read from proxy issues
This patch provide a better protocol handling than the way it was done.
It expects to read buffer length to then determine if more reading is
necessary. At last, it retrieves the file descriptor passed as an out
of band data.

However, there is still an unconsistent behavior related to the OOB data
received. Sometimes when recvmsg() is called to fast, it cannot receive
OOB data containing the fd, ending in a MSG_CTRUNC error. This issue is
only related to OOB data because this error means that nothing is sent
as OOB data. In the case of a regular buffer passed on the band, even if
it is truncated, you can still retrieve the other part by issuing another
recvmsg() call.
That's why we need to make sure to wait a bit before we can try to read
something on the socket, avoiding OOB corruption.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>